Bot/requirements.txt
aiogram
Pillow
torch
torchvision
numpy



